Factors yaml constructors out into separate module
